Fixed the Makefile, and attempted a docker-compose solution
To resolve the problem I had with make I found an older version of make for windows, installed it, and added the path to my environment variables.

My docker-compose solution appears to work: it will bundle the image, but does not execute the second command after the &&. It simply terminates with state: exit 0.